requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
google-cloud-firestore